refactor: create goal module, move relevant items to it
