+++ src/main.rs
@@ -7,6 +7,7 @@ mod control_display;
 mod event_scheduler;
 mod from_component;
 mod gameplay;
+mod goal;
 mod gravestone;
 mod history;
 mod level_select;
@@ -86,7 +87,7 @@ fn main() {
         .add_plugin(level_transition::LevelTransitionPlugin)
         .add_event::<history::HistoryCommands>()
         .add_event::<gameplay::DeathEvent>()
-        .add_event::<gameplay::GoalEvent>()
+        .add_event::<goal::GoalEvent>()
         .insert_resource(LdtkSettings {
             set_clear_color: SetClearColor::FromEditorBackground,
             ..default()
@@ -94,7 +95,7 @@ fn main() {
         .insert_resource(Msaa { samples: 1 })
         .insert_resource(level_selection.clone())
         .insert_resource(level_transition::TransitionTo(level_selection))
-        .insert_resource(resources::GoalGhostSettings::NORMAL)
+        .insert_resource(goal::GoalGhostSettings::NORMAL)
         .insert_resource(resources::RewindSettings::NORMAL)
         .insert_resource(resources::PlayZonePortion(0.75))
         .add_startup_system(gameplay::transitions::spawn_camera)
@@ -109,9 +110,7 @@ fn main() {
                 .run_not_in_state(GameState::AssetLoading)
                 .run_on_event::<bevy::window::WindowResized>(),
         )
-        .add_system(
-            gameplay::transitions::spawn_goal_ghosts.run_in_state(GameState::LevelTransition),
-        )
+        .add_system(goal::spawn_goal_ghosts.run_in_state(GameState::LevelTransition))
         .add_system(
             gameplay::systems::check_death
                 .run_in_state(GameState::Gameplay)
@@ -124,15 +123,15 @@ fn main() {
                 .label(history::FlushHistoryCommands),
         )
         .add_system(
-            gameplay::systems::check_goal
+            goal::check_goal
                 .run_in_state(GameState::Gameplay)
                 .after(SystemLabels::CheckDeath),
         )
         .add_system(gameplay::transitions::spawn_death_card.run_in_state(GameState::Gameplay))
-        .add_system(sugar::goal_ghost_animation.run_not_in_state(GameState::AssetLoading))
-        .add_system(sugar::goal_ghost_event_sugar.run_not_in_state(GameState::AssetLoading))
+        .add_system(goal::goal_ghost_animation.run_not_in_state(GameState::AssetLoading))
+        .add_system(goal::goal_ghost_event_sugar.run_not_in_state(GameState::AssetLoading))
         .add_system(sugar::animate_grass_system.run_not_in_state(GameState::AssetLoading))
-        .register_ldtk_entity::<bundles::GoalBundle>("Goal")
+        .register_ldtk_entity::<goal::GoalBundle>("Goal")
         .register_ldtk_entity::<bundles::GrassBundle>("Grass")
         .register_ldtk_int_cell::<bundles::ExorcismBlockBundle>(2);
